+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/movimentacao/motivos/_model/MotivoMovimentoListagem.java
@@ -1,0 +1,23 @@
+package br.com.zalf.prolog.webservice.frota.pneu.movimentacao.motivos._model;
+
+import lombok.Data;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+/**
+ * Created on 2020-04-01
+ *
+ * @author Gustavo Navarro (https://github.com/gustavocnp95)
+ */
+@Data
+public final class MotivoMovimentoListagem {
+
+    @NotNull
+    private final Long codMotivoMovimento;
+    @NotNull
+    private final String descricaoMotivoMovimento;
+    @Nullable
+    private final String codAuxiliar;
+    private final boolean ativo;
+
+}