+++ src/main/java/br/com/zalf/prolog/webservice/frota/socorrorota/_model/SocorroRotaAtendimentoVisualizacao.java
@@ -1,7 +1,10 @@
 package br.com.zalf.prolog.webservice.frota.socorrorota._model;
 
+import com.google.gson.annotations.SerializedName;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+
+import java.time.Duration;
 import java.time.LocalDateTime;
 
 /**
@@ -28,6 +31,17 @@ public final class SocorroRotaAtendimentoVisualizacao {
     private final String modeloAparelhoAtendimentoSocorro;
     @Nullable
     private final String imeiAparelhoAtendimentoSocorro;
+    @Nullable
+    private final LocalDateTime dataHoraDeslocamentoInicio;
+    @Nullable
+    private final LocalizacaoSocorroRota localizacaoDescolamentoInicio;
+    @Nullable
+    private final LocalDateTime dataHoraDeslocamentoFim;
+    @Nullable
+    private final LocalizacaoSocorroRota localizacaoDescolamentoFim;
+    @Nullable
+    @SerializedName("tempoAberturaAtendimentoEmSegundos")
+    private final Duration tempoAberturaAtendimento;
 
     public SocorroRotaAtendimentoVisualizacao(@NotNull final Long codColaboradorResponsavelAtendimento,
                                               @NotNull final String nomeResponsavelAtendimentoSocorro,
@@ -37,7 +51,12 @@ public final class SocorroRotaAtendimentoVisualizacao {
                                               @Nullable final String enderecoAutomaticoAtendimentoSocorro,
                                               @Nullable final String marcaAparelhoAtendimentoSocorro,
                                               @Nullable final String modeloAparelhoAtendimentoSocorro,
-                                              @Nullable final String imeiAparelhoAtendimentoSocorro) {
+                                              @Nullable final String imeiAparelhoAtendimentoSocorro,
+                                              @Nullable final LocalDateTime dataHoraDeslocamentoInicio,
+                                              @Nullable final LocalizacaoSocorroRota localizacaoDescolamentoInicio,
+                                              @Nullable final LocalDateTime dataHoraDeslocamentoFim,
+                                              @Nullable final LocalizacaoSocorroRota localizacaoDescolamentoFim,
+                                              @Nullable final Duration tempoAberturaAtendimento) {
         this.codColaboradorResponsavelAtendimento = codColaboradorResponsavelAtendimento;
         this.nomeResponsavelAtendimentoSocorro = nomeResponsavelAtendimentoSocorro;
         this.observacaoAtendimento = observacaoAtendimento;
@@ -47,6 +66,11 @@ public final class SocorroRotaAtendimentoVisualizacao {
         this.marcaAparelhoAtendimentoSocorro = marcaAparelhoAtendimentoSocorro;
         this.modeloAparelhoAtendimentoSocorro = modeloAparelhoAtendimentoSocorro;
         this.imeiAparelhoAtendimentoSocorro = imeiAparelhoAtendimentoSocorro;
+        this.dataHoraDeslocamentoInicio = dataHoraDeslocamentoInicio;
+        this.localizacaoDescolamentoInicio = localizacaoDescolamentoInicio;
+        this.dataHoraDeslocamentoFim = dataHoraDeslocamentoFim;
+        this.localizacaoDescolamentoFim = localizacaoDescolamentoFim;
+        this.tempoAberturaAtendimento = tempoAberturaAtendimento;
     }
 
     @NotNull
@@ -78,5 +102,26 @@ public final class SocorroRotaAtendimentoVisualizacao {
     @NotNull
     public StatusSocorroRota getStatusReferencia() {
         return StatusSocorroRota.EM_ATENDIMENTO;
+    }
+
+    @Nullable
+    public LocalDateTime getDataHoraDeslocamentoInicio() { return dataHoraDeslocamentoInicio; }
+
+    @Nullable
+    public LocalizacaoSocorroRota getLocalizacaoDescolamentoInicio() {
+        return localizacaoDescolamentoInicio;
+    }
+
+    @Nullable
+    public LocalDateTime getDataHoraDeslocamentoFim() { return dataHoraDeslocamentoFim; }
+
+    @Nullable
+    public LocalizacaoSocorroRota getLocalizacaoDescolamentoFim() {
+        return localizacaoDescolamentoFim;
+    }
+
+    @Nullable
+    public Duration getTempoAberturaAtendimento() {
+        return tempoAberturaAtendimento;
     }
 }