+++ src/main/java/br/com/zalf/prolog/webservice/integracao/praxio/data/SistemaGlobusPiccoloturDao.java
@@ -5,6 +5,7 @@ import br.com.zalf.prolog.webservice.integracao.praxio.ordensservicos.model.Chec
 import org.jetbrains.annotations.NotNull;
 
 import java.sql.Connection;
+import java.util.List;
 
 /**
  * Created on 17/06/19.
@@ -99,4 +100,14 @@ public interface SistemaGlobusPiccoloturDao {
      */
     boolean verificaModeloChecklistIntegrado(@NotNull final Long codUnidade,
                                              @NotNull final Long codModeloChecklist) throws Throwable;
+
+    /**
+     * Método responsável por verificar se algum dos itens prestes a serem resolvidos é integrado. Não permitimos o
+     * fechamento de itens integrados. O fechamento deve ocorrer apenas via integração.
+     *
+     * @param codItensResolver Lista com os códigos de itens a serem validados.
+     * @return <code>TRUE</code> caso algum dos itens é integrado, <code>FALSE</code> caso contrátio.
+     * @throws Throwable Se algum erro acontecer
+     */
+    boolean verificaItensIntegrados(@NotNull final List<Long> codItensResolver) throws Throwable;
 }