+++ src/main/java/br/com/zalf/prolog/webservice/integracao/praxio/movimentacao/GlobusPiccoloturLocalMovimento.java
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName;
 import lombok.Data;
 import org.jetbrains.annotations.NotNull;
 
-import static br.com.zalf.prolog.webservice.integracao.praxio.GlobusPiccoloturConstants.COD_UNIDADE_NOME_LOCAL_SEPARATOR;
+import static br.com.zalf.prolog.webservice.integracao.praxio.GlobusPiccoloturConstants.COD_UNIDADE_NOME_LOCAL_MOVIMENTO_SEPARATOR;
 
 /**
  * Created on 2020-03-28
@@ -25,6 +25,9 @@ public final class GlobusPiccoloturLocalMovimento {
 
     @NotNull
     public String getLocalAsOpcaoSelecao() {
-        return this.codUnidadeProlog.toString().concat(COD_UNIDADE_NOME_LOCAL_SEPARATOR).concat(this.nomeLocalGlobus);
+        return this.codUnidadeProlog
+                .toString()
+                .concat(COD_UNIDADE_NOME_LOCAL_MOVIMENTO_SEPARATOR)
+                .concat(this.nomeLocalGlobus);
     }
 }