+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 Change Log
 ==========
+<a name="v1.0.26"></a>
+## Version [v1.0.26](https://github.com/luizfp/PrologWebService/compare/v1.0.25...v1.0.26) (2020-04-12)
+
+### Features
+* Cria dashboard de socorros por status nos ultimos 30 dias (PL-2618)
+* Cria estrutura para cadastro e configuração de motivos para o processo de movimentação (PL-2607)
+* Permite alteração da pressão recomendada dos pneus (PL-2570)
+
+### Refactors
+* Permitir fechamento de O.Ss originadas no Prolog (PLI-99)
+* Adiciona os registros de deslocamento para socorros em rota (PL-2631)
+* Adiciona informações de tempo entre cada status do socorro em rota (PL-2585)
+* Roteia método de sincronia de checklist offline (PLI-118)
+
+### Bug Fixes
+* Corrige parse sem replace de virgula no import de csv (PL-2573)
+
 <a name="v1.0.25"></a>
 ## Version [v1.0.25](https://github.com/luizfp/PrologWebService/compare/v1.0.24...v1.0.25) (2020-04-02)